pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tensorpack-f"
version = "0.1.8"
description = "TensorPack - tools for tensor/matrix conversion and semantic connections"
readme = "README.md"
requires-python = "==3.9.*"
authors = [ { name="TensorPack Authors" } ]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Operating System :: OS Independent",
]

dependencies = [
	"numpy",
	"pandas",
	"scikit-learn",
	"scipy",
	"matplotlib",
	"seaborn",
	"pillow",
	"opencv-python-headless",
	"pyarrow",
	"h5py",
	"netCDF4",
	"tifffile",
	"nibabel",
	"zarr",
	"requests",
	"cryptography",
	
	"numba",
	"joblib",
	"networkx",
	"tqdm",
	"openpyxl",
	"xlwt",
	"torch",
	"transformers",
	"jsonschema",
	"sqlalchemy",
	"psutil",
	"python-dateutil",
	"pandas-flavor",
	"pyyaml",
	"lxml",
	"scikit-image",
	"fastparquet",
	"regex",
	"fonttools",
	"Flask",
	"rich",
	"PyJWT",
]

[project.scripts]
tensorpack = "tensorpack.script:main"

[tool.setuptools]
include-package-data = false
zip-safe = false

[tool.setuptools.packages]
find = {}

[tool.setuptools.package-data]
"*" = ["*.pyc"]
"tensorpack" = ["__pycache__/*.pyc"]